move _add_constant out
